modify: logging, refactor delete controller

+++ controllers/book.controller.js
@@ -1,4 +1,5 @@
 import { Book } from "../models/Books.js"
+import { logger } from "../utils/jwt_utils.js"
 
 export class BookController {
 
@@ -6,19 +7,23 @@ export class BookController {
         const page = req.query.page || 1
         const limit = 4
         const skip = (page - 1) * limit
-
+        // query book by username
         const username = req.query.username;
         const query = (req.query.username !== undefined) ? {uploaded_by:username} : {}
 
+        // TODO Query based on the presence of other parameters
+
         try {
             const books = await Book.find(query).sort({createdAt: "desc"}).skip(skip).limit(limit);
+            logger.info("All books fetched!")
+
             return res.status(200).json({
                 message: "Books fetched",
                 status: true,
                 data: books
-            })
+            }) 
         } catch (error) {
-            console.log(error);
+            logger.error(error);
             return res.status(400).json({
                 message: error.message,
                 status: false,
@@ -28,22 +33,22 @@ export class BookController {
     }
  
     async add_book(req, res) {
-        console.log(req.session.username);
         try {
-            console.log(req.body);
-            const uploaded_by = req.session.username
-
-            console.log(`Book uploaded by ${uploaded_by}`);
+            logger.info(JSON.stringify(req.body)); 
+            const uploaded_by = req.user.username
             
             const book = new Book({...req.body, uploaded_by});
-            
             const newBook = await book.save()
+
+            logger.info(`${book.title} uploaded by ${uploaded_by}`)
+
             return res.status(201).json({
                 message: "Book added!",
                 status: true,
                 data: newBook
             })            
         } catch (error) {
+            logger.error(`Add Book error - ${error.message}`)
             return res.status(400).json({
                 message: error.message,
                 status: false,
@@ -53,14 +58,17 @@ export class BookController {
     }
 
     async get_book(req, res) {
+        const id = req.params.id;
+
         try {
-            const id = req.params.id;
-            const book = await Book.findById(id)
+            const book = await Book.findById(id);
+            logger.info(`${book.title} fetched!`);
+
             return res.status(200).json({
                 message: "Book fetched",
                 status: true,
                 data: book
-            })
+            }) 
         } catch (error) {
             return res.status(400).json({
                 message: error.message,
@@ -71,15 +79,25 @@ export class BookController {
     }
 
     async delete_book(req,res) { 
+        const id = req.params.id;
+        
         try {
-            const id = req.params.id;
+            // does the book exist
+            const book = await Book.findById({_id : id});
+            if (!book) {
+                return res.status(400).json({message: "Book not found!"})
+            } 
+            //delete the book
             await Book.findByIdAndDelete(id);
+            logger.info(`${book.title} deleted by ${req.user.username}`)
+
             return res.status(200).json({
                 message: "Book deleted",
                 status: true,
                 data: null
             })
         } catch (error) {
+            logger.error(`Error deleting book: ${error.message}`)
             return res.status(400).json({
                 message: error.message,
                 status: false,
